requirement.txt
Pillow==9.5.0
playsound==1.3.0
